+++ Engine/Menu/MainMenu.cpp
@@ -549,11 +549,11 @@ void MainMenu::Update()
         {
             totem[i].setPosition((float(360) + float(918) * (i)) + g_x[3] / 1.4f, 2160);
 
-            if (mouseCtrl->getMousePos().x / (window->getSize().x / 3840.f) > totem[i].getPosition().x)
+            if (mouseCtrl->getMousePos().x / (static_cast<float>(window->getSize().x) / 3840.f) > totem[i].getPosition().x)
             {
-                if (mouseCtrl->getMousePos().x / (window->getSize().x / 3840.f) < (totem[i].getPosition().x + totem[i].getTransformedBounds().size.x))
+                if (mouseCtrl->getMousePos().x / (static_cast<float>(window->getSize().x) / 3840.f) < (totem[i].getPosition().x + totem[i].getTransformedBounds().size.x))
                 {
-                    if (mouseCtrl->getMousePos().y / (window->getSize().y / 2160.f) > totem[i].getPosition().y - totem[i].getTransformedBounds().size.y)
+                    if (mouseCtrl->getMousePos().y / (static_cast<float>(window->getSize().y) / 2160.f) > totem[i].getPosition().y - totem[i].getTransformedBounds().size.y)
                     {
                         totem_sel = i;
                         mouseInBounds = true;
