+++ Engine/Mission/Units/HitboxFrame.cpp
@@ -13,7 +13,7 @@ HitboxFrame::~HitboxFrame()
 
 float HitboxFrame::minProjection(float axisAngle, float object_x, float object_y)
 {
-    float projectionLength = 99999999;
+    float projectionLength = 99999999.f;
     std::vector<sf::Vector2f> currentVertices = getCurrentVertices();
 
     //cout<<" - SEARCHING FOR MIN PROJECTION"<<endl;
@@ -95,7 +95,7 @@ void HitboxFrame::calcMaxWidth(float maxWidth, bool forceRecalc)
     for (int i = 1; i < vertices.size(); i++) 
     {
         sf::Vector2f currentVertex = vertices[i];
-        float dist = sqrt(pow(currentVertex.x - firstVertex.x, 2) + pow(currentVertex.y - firstVertex.y, 2) * 1.0);
+        float dist = sqrtf(pow(currentVertex.x - firstVertex.x, 2) + pow(currentVertex.y - firstVertex.y, 2) * 1.0);
         if (dist>maxlength) 
         {
             maxlength = dist;
@@ -109,7 +109,7 @@ void HitboxFrame::calcMaxWidth(float maxWidth, bool forceRecalc)
     for (int i = 1; i < vertices.size(); i++)
     {
         sf::Vector2f currentVertex = vertices[i];
-        float dist = sqrt(pow(currentVertex.x - 0, 2) + pow(currentVertex.y - 0, 2) * 1.0);
+        float dist = sqrtf(pow(currentVertex.x - 0, 2) + pow(currentVertex.y - 0, 2) * 1.0);
         if (dist > maxlength_origin)
         {
             maxlength_origin = dist;
